# [3/3] Return errors immediately instead of using goto