--- a/src/haswell/dom_parser_implementation.cpp
+++ b/src/haswell/dom_parser_implementation.cpp
@@ -107,13 +107,12 @@ WARN_UNUSED bool implementation::validate_utf8(const char *buf, size_t len) cons
 }
 
 WARN_UNUSED error_code dom_parser_implementation::stage2(dom::document &_doc) noexcept {
-  error_code result = stage2::parse_structurals<false>(*this, _doc);
-  if (result) { return result; }
+  if (auto error = stage2::parse_structurals<false>(*this, _doc)) { return error; }
 
   // If we didn't make it to the end, it's an error
   if ( next_structural_index != n_structural_indexes ) {
     logger::log_string("More than one JSON value at the root of the document, or extra characters at the end of the JSON!");
-    return error = TAPE_ERROR;
+    return TAPE_ERROR;
   }
 
   return SUCCESS;
@@ -124,8 +123,8 @@ WARN_UNUSED error_code dom_parser_implementation::stage2_next(dom::document &_do
 }
 
 WARN_UNUSED error_code dom_parser_implementation::parse(const uint8_t *_buf, size_t _len, dom::document &_doc) noexcept {
-  error_code err = stage1(_buf, _len, false);
-  if (err) { return err; }
+  auto error = stage1(_buf, _len, false);
+  if (error) { return error; }
   return stage2(_doc);
 }