@lemire 2020-07-30
At a glance, I am not preoccupied too much by the performance effect and I do agree with the general strategy regarding performance and stage 2.  My take has never been "don't allow any regression"... but rather "measure and track the effect of your changes".