@lemire 2020-07-30
I'll comment here about the performance issue instead of commenting on each distinct PR.

What these PRs do is to *reduce* (generally) the instruction count, but to *increase the number of cycles*.  This results in a significant performance degradation. The slowdown is real. For example, we may end up 7% slower on twitter.json in stage 2.

But, importantly, this is *not* caused by an increase in the instruction count.

What is happening? It could be the result of more branch misses, but no. That is not the case. We don't have more branch misses, apparently.

So we probably suffer from less superscalarity. There may be longer dependency chains, or something like it.

Everything else being equal, we would reject these PR, since they result in lower performance. But... 

1. We have lots of experience tweaking stage 2 and we know that seemingly inconsequential changes may result in wild performance variations. If we don't give ourselves a good margin, we will never be able to make substantial changes.
2. These changes open the door (very carefully) to more versatile designs. It is necessary for us to reach version 1.0. We have no choice.
3. These changes (arguably) simplify the code.
4. We are still at about 3 instructions per cycle in both stage 1 and stage 2. I'd be more nervous about dropping below 3 instructions per cycle.