@jkeiser 2020-08-02
@lemire after rebasing against master, this is has a wildly unpredictable performance impact (these are *overall* numbers, by the way, not just stage 2), including a hit on twitter.json (gcc10 skylake-x):

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| mesh                             |  11306 | 311.0 | 272.5 |   12% | 978.9 | 986.9 |    0% |  10152 |   9784 |    3% |      3 |      9 | -200% |  60498 |  60832 |    0% |
| marine_ik                        |  46616 | 298.3 | 268.5 |   10% | 912.6 | 903.9 |    0% |  53695 |  53082 |    1% | 111046 | 111469 |    0% | 255885 | 255877 |    0% |
| numbers                          |   2345 | 253.7 | 231.2 |    8% | 790.8 | 790.8 |    0% |   2190 |   2170 |    0% |      0 |      0 |    0% |      0 |      0 |    0% |
| mesh.pretty                      |  24646 | 179.8 | 166.0 |    7% | 579.2 | 579.5 |    0% |  13553 |  13777 |   -1% |    898 |    672 |   25% |  88147 |  88167 |    0% |
| canada                           |  35172 | 289.1 | 273.1 |    5% | 982.2 | 980.9 |    0% |  18546 |  19699 |   -6% |  23380 |  22607 |    3% | 154761 | 154788 |    0% |
| citm_catalog                     |  26987 |  84.9 |  83.5 |    1% | 314.5 | 307.0 |    2% |   1752 |   1732 |    1% |   1283 |   1156 |    9% |  89467 |  89478 |    0% |
| gsoc-2018                        |  51997 |  73.4 |  75.3 |   -2% | 169.8 | 169.1 |    0% |  30913 |  31988 |   -3% |  53116 |  48815 |    8% | 168469 | 168456 |    0% |
| apache_builds                    |   1988 |  87.6 |  89.9 |   -2% | 311.4 | 309.6 |    0% |     97 |    104 |   -7% |      0 |      0 |    0% |      2 |      0 |    0% |
| instruments                      |   3442 | 108.8 | 112.1 |   -3% | 373.6 | 369.2 |    1% |    473 |    480 |   -1% |     54 |      2 |   96% |   1784 |    503 |   71% |
| twitterescaped                   |   8787 | 191.0 | 197.1 |   -3% | 515.5 | 512.6 |    0% |   4952 |   5217 |   -5% |      0 |      3 |    0% |  30813 |  29913 |    2% |
| github_events                    |   1017 |  78.9 |  81.7 |   -3% | 269.4 | 267.0 |    0% |    102 |    123 |  -20% |      0 |      0 |    0% |      0 |      0 |    0% |
| twitter                          |   9867 |  92.9 |  96.6 |   -3% | 296.8 | 294.2 |    0% |   3381 |   3470 |   -2% |     27 |      9 |   66% |  32722 |  33478 |   -2% |
| random                           |   7976 | 145.4 | 151.4 |   -4% | 497.3 | 494.5 |    0% |   2066 |   2034 |    1% |    167 |      8 |   95% |  37461 |  37310 |    0% |
| update-center                    |   8330 | 116.7 | 122.8 |   -5% | 344.8 | 343.5 |    0% |   5928 |   5692 |    3% |      2 |      5 | -150% |  32764 |  32912 |    0% |

I tried removing the `unlikely` modifiers on the theory that they might interact differently with `return` than `goto`, but they either didn't help, or made things universally worse.

I am not entirely convinced unlikely is helping us in stage 2. I suspect that anything that moves code around has massive random performance impacts, and that if we make enough changes we'll settle on something that looks good each time, but really has no sound basis or may even objectively make things worse. i.e. it happens to find a binary format niche that is better performance, but it actually depresses performance in all cases, and other less impactful random changes could have found the same niche with even better throughput.

I'm going to merge, regardless of the impact to twitter, but we should keep an eye on this.