@lemire 2020-08-03
> I am not entirely convinced unlikely is helping us in stage 2.

I agree that it is flaky, but I am not sure we should just ignore the problem. I submit to you that, maybe, we just have not developed the proper strategy. If you look at the assembly for something like the colon check, if you do not use 'unlikely' (to check for the absence of a colon), you get strange assembly that is clearly not optimized for the fact that the colon is almost always present.