@jkeiser 2020-08-03
I don't think we should ignore it either :) But I'll say this: I made some changes last night and got negative performance. Removing these unlikely statements improved it.

Something is going on, I just suspect the need for unlikely is a symptom.