@lemire 2020-08-03
@jkeiser Compared to v0.4.7, our main branch is now about 5% slower. So there is room for re-optimization. This can be revisited later.